@@ -9,30 +9,27 @@ Bundler.require(*Rails.groups)
 module CommunityPartnerPlatform
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.0
+    config.load_defaults 5.2
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded after loading
     # the framework and any gems in your application.
 
-    %w(
-      lib
-      app/models/community_program_quality_element
-      app/models/school
-    ).each do |path|
-      config.autoload_paths << Rails.root.join(path)
-      config.eager_load_paths << Rails.root.join(path)
-    end
+    #
+    # App specific changes
+    #
 
-    # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-    # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
+    # %w(lib).each do |path|
+    #   config.autoload_paths << Rails.root.join(path)
+    #   config.eager_load_paths << Rails.root.join(path)
+    # end
+
     config.time_zone = 'Pacific Time (US & Canada)'
 
-    # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
-    # config.i18n.default_locale = :de
-    config.i18n.load_path = Dir[Rails.root.join('config', 'locales', '**', '*.{rb,yml}')] - 
+    # modify locale files loaded to only include the overrides for the
+    # configured subdomain
+    config.i18n.load_path = Dir[Rails.root.join('config', 'locales', '**', '*.{rb,yml}')] -
       Dir[Rails.root.join('config', 'locales', 'overrides', '**', '*.{rb,yml}')] +
       Dir[Rails.root.join('config', 'locales', 'overrides', Settings.application.subdomain, '*.{rb,yml}')]
   end
