@@ -8,11 +8,13 @@ Rails.application.config.assets.version = '1.1'
 # Add Yarn node_modules folder to the asset load path.
 Rails.application.config.assets.paths << Rails.root.join('node_modules')
 
-Rails.application.config.assets.paths << Rails.root.join('app/assets/fonts')
-
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in the app/assets
 # folder are already added.
+# Rails.application.config.assets.precompile += %w( admin.js admin.css )
+
+Rails.application.config.assets.paths << Rails.root.join('app/assets/fonts')
+
 Rails.application.config.assets.precompile += %w(
   *.png *.jpg *.jpeg *.gif
   *.eot *.svg *.ttf *.woff
